preparativos para o evento do dia 12

+++ commands/utility/top.py
@@ -36,7 +36,8 @@ class TopClass(commands.Cog):
                                 f"{self.st[67]} `top rpg` Top 20 dos usuarios com maiores LEVEIS do RPG.\n"
                                 f"{self.st[67]} `top raid` Top 20 dos usuarios com maiores RAIDS do RPG.\n"
                                 f"{self.st[67]} `top fragment` Top 20 dos usuarios com mais FRAGMENTOS.\n"
-                                f"{self.st[67]} `top blessed` Top 20 dos usuarios com mais BLESSEDS.")
+                                f"{self.st[67]} `top blessed` Top 20 dos usuarios com mais BLESSEDS.\n"
+                                f"{self.st[67]} `top event` Top 20 das guildas com mais pontos de EVENTO.")
             top.set_author(name=ctx.author.name, icon_url=ctx.author.avatar_url)
             top.set_thumbnail(url=self.bot.user.avatar_url)
             top.set_footer(text="Ashley ® Todos os direitos reservados.")
@@ -186,6 +187,21 @@ class TopClass(commands.Cog):
                              "**mesmo que demore, aguarde o fim do processamento...**")
         top = await self.bot.data.get_rank_blessed(20, ctx)
         await ctx.send(f'<:rank:519896825411665930>|**TOP BLESSED**```py\n{top}```')
+        await msg.delete()
+
+    @check_it(no_pm=True)
+    @commands.cooldown(1, 5.0, commands.BucketType.user)
+    @commands.check(lambda ctx: Database.is_registered(ctx, ctx))
+    @top.group(name='event')
+    async def _event(self, ctx):
+        if not self.bot.event_special and ctx.author.id not in self.bot.staff:
+            return await ctx.send(f"<:negate:721581573396496464>│`ATUALMENTE NAO TEM NENHUM EVENTO ESPECIAL!`")
+
+        msg = await ctx.send("<a:loading:520418506567843860>│ `AGUARDE, ESTOU PROCESSANDO SEU PEDIDO!`\n"
+                             "**mesmo que demore, aguarde o fim do processamento...**")
+
+        top = await self.bot.data.get_rank_event(20, ctx)
+        await ctx.send(f'<:rank:519896825411665930>|**TOP EVENT**```py\n{top}```')
         await msg.delete()
